Rename package majorversions to versionbumps
