+++ operator/bundle_helpers/pkg/csv/version.go
@@ -7,7 +7,7 @@ import (
 	"slices"
 	"strconv"
 
-	"github.com/stackrox/rox/pkg/version/majorversions"
+	"github.com/stackrox/rox/pkg/version/versionbumps"
 )
 
 // XyzVersion represents a semantic version with major.minor.patch components
@@ -88,7 +88,7 @@ func (v XyzVersion) IsEqualOrBefore(other XyzVersion) bool {
 // GetPreviousYStream returns the previous Y-Stream version.
 // Y-Stream versions have patch number = 0 (e.g., 3.73.0, 3.74.0, 4.0.0)
 func GetPreviousYStream(v XyzVersion) (*XyzVersion, error) {
-	prev, err := majorversions.GetPreviousYStream(majorversions.XYVersion{X: v.X, Y: v.Y})
+	prev, err := versionbumps.GetPreviousYStream(versionbumps.XYVersion{X: v.X, Y: v.Y})
 	if err != nil {
 		return nil, err
 	}
