+++ pkg/version/versionbumps/version_bumps.go
@@ -1,4 +1,4 @@
-package majorversions
+package versionbumps
 
 import (
 	"cmp"
